requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
motor==3.3.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0
certifi==2023.11.17
